+++ perf/_timeit.py
@@ -2,19 +2,114 @@ from __future__ import division, print_function, absolute_import
 
 import itertools
 import sys
-import timeit
 import traceback
 
 import perf
 
 
-def sample_func(loops, timer):
-    if perf.python_implementation() == 'pypy':
-        inner = timer.make_inner()
-        return inner(loops, timer.timer)
-    else:
-        it = itertools.repeat(None, loops)
-        return timer.inner(it, timer.timer)
+PYPY = (perf.python_implementation() == 'pypy')
+DUMMY_SRC_NAME = "<timeit-src>"
+
+# Don't change the indentation of the template; the reindent() calls
+# in Timer.__init__() depend on setup being indented 4 spaces and stmt
+# being indented 8 spaces.
+TEMPLATE = """
+def inner(_it, _timer{init}):
+    {setup}
+    _t0 = _timer()
+    for _i in _it:
+        {stmt}
+    _t1 = _timer()
+    return _t1 - _t0
+"""
+
+PYPY_TEMPLATE = """
+def inner(_it, _timer{init}):
+    {setup}
+    _t0 = _timer()
+    while _it > 0:
+        _it -= 1
+        {stmt}
+    _t1 = _timer()
+    return _t1 - _t0
+"""
+
+
+def reindent(src, indent):
+    return src.replace("\n", "\n" + " " * indent)
+
+
+class Timer:
+    def __init__(self, stmt="pass", setup="pass",
+                 globals=None):
+        self.local_ns = {}
+        self.global_ns = {} if globals is None else globals
+        self.filename = DUMMY_SRC_NAME
+
+        init = ''
+        if isinstance(setup, str):
+            # Check that the code can be compiled outside a function
+            compile(setup, self.filename, "exec")
+            stmtprefix = setup + '\n'
+            setup = reindent(setup, 4)
+        elif callable(setup):
+            self.local_ns['_setup'] = setup
+            init += ', _setup=_setup'
+            stmtprefix = ''
+            setup = '_setup()'
+        else:
+            raise ValueError("setup is neither a string nor callable")
+
+        if isinstance(stmt, str):
+            # Check that the code can be compiled outside a function
+            compile(stmtprefix + stmt, self.filename, "exec")
+            stmt = reindent(stmt, 8)
+        elif callable(stmt):
+            self.local_ns['_stmt'] = stmt
+            init += ', _stmt=_stmt'
+            stmt = '_stmt()'
+        else:
+            raise ValueError("stmt is neither a string nor callable")
+
+        if PYPY:
+            template = PYPY_TEMPLATE
+        else:
+            template = TEMPLATE
+        src = template.format(stmt=stmt, setup=setup, init=init)
+        self.src = src  # Save for traceback display
+
+    def make_inner(self):
+        # PyPy tweak: recompile the source code each time before
+        # calling inner(). There are situations like Issue #1776
+        # where PyPy tries to reuse the JIT code from before,
+        # but that's not going to work: the first thing the
+        # function does is the "-s" statement, which may declare
+        # new classes (here a namedtuple). We end up with
+        # bridges from the inner loop; more and more of them
+        # every time we call inner().
+        code = compile(self.src, self.filename, "exec")
+        global_ns = dict(self.global_ns)
+        local_ns = dict(self.local_ns)
+        exec(code, global_ns, local_ns)
+        return local_ns["inner"]
+
+    def update_linecache(self, file=None):
+        import linecache
+
+        linecache.cache[self.filename] = (len(self.src),
+                                          None,
+                                          self.src.split("\n"),
+                                          self.filename)
+
+    def sample_func(self, loops):
+        inner = self.make_inner()
+        timer = perf.perf_counter
+        if not PYPY:
+            it = itertools.repeat(None, loops)
+            return inner(it, timer)
+        else:
+            # PyPy
+            return inner(loops, timer)
 
 
 def strip_statements(statements):
@@ -40,9 +135,7 @@ def create_timer(stmt, setup, globals):
     stmt = "\n".join(stmt)
     setup = "\n".join(setup)
 
-    return timeit.Timer(stmt, setup,
-                        timer=perf.perf_counter,
-                        globals=globals)
+    return Timer(stmt, setup, globals=globals)
 
 
 def display_error(timer, stmt, setup):
@@ -61,9 +154,9 @@ def display_error(timer, stmt, setup):
         print()
 
     if timer is not None:
-        timer.print_exc()
-    else:
-        traceback.print_exc()
+        timer.update_linecache()
+
+    traceback.print_exc()
 
 
 def bench_timeit(runner, name, stmt, setup,
@@ -84,8 +177,11 @@ def bench_timeit(runner, name, stmt, setup,
     metadata = {}
     if func_metadata:
         metadata.update(func_metadata)
-    metadata['timeit_setup'] = format_statements(setup)
+    if setup:
+        metadata['timeit_setup'] = format_statements(setup)
     metadata['timeit_stmt'] = format_statements(stmt)
+
+    orig_stmt = stmt
 
     # args must not be modified, it's passed to the worker process,
     # so use local variables.
@@ -104,10 +200,9 @@ def bench_timeit(runner, name, stmt, setup,
     timer = None
     try:
         timer = create_timer(stmt, setup, globals)
-        runner.bench_sample_func(name, sample_func,
-                                 timer, **kwargs)
+        runner.bench_sample_func(name, timer.sample_func, **kwargs)
     except SystemExit:
         raise
     except:
-        display_error(timer, stmt, setup)
+        display_error(timer, orig_stmt, setup)
         sys.exit(1)